fake-news-ai-platform/backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
pydantic==1.10.9
transformers==4.34.0
torch==2.0.1
requests==2.31.0
python-jose==3.3.0
celery==5.3.1
redis==4.5.5
openai==0.27.6
pytest==7.4.0
httpx==0.24.0
pytest-mock==3.11.1
